requirements.txt
Django
djangorestframework
openai
PyPDF2
python-dotenv
